# [PERF] PromQL: use memoized metric hashes as cache keys in VectorBinop