This is the fourth PR in a series of PromQL join optimizations that I am planning. It depends on #17148, please review that first.

The original implementation was using a concatenation (which had to be built for every resulting sample) of left and right metric label sets as a string key into result cache. With this change, we will use a pair of (left, right) metrics hashes for this cache key instead. Hashes will be computed and memoized lazily, only once per input time series. 

Benchmark results (relative to #17148):
```
                                                                                                                │ after3.txt  │              after4.txt              │
                                                                                                                │   sec/op    │    sec/op     vs base                │
JoinQuery/expr=rpc_request_success_total_+_rpc_request_error_total,steps=10000-10                                  3.512 ± 3%     2.545 ± 2%  -27.54% (p=0.000 n=10)
JoinQuery/expr=rpc_request_success_total_+_ON_(job,_instance)_GROUP_LEFT_rpc_request_error_total,steps=10000-10    4.010 ± 3%     3.254 ± 4%  -18.87% (p=0.000 n=10)
JoinQuery/expr=rpc_request_success_total_AND_rpc_request_error_total{instance=~"0.*"},steps=10000-10              593.1m ± 3%    606.0m ± 3%   +2.17% (p=0.019 n=10)
JoinQuery/expr=rpc_request_success_total_OR_rpc_request_error_total{instance=~"0.*"},steps=10000-10                1.051 ± 3%     1.036 ± 3%        ~ (p=0.247 n=10)
JoinQuery/expr=rpc_request_success_total_UNLESS_rpc_request_error_total{instance=~"0.*"},steps=10000-10           987.5m ± 2%   1025.8m ± 2%   +3.88% (p=0.004 n=10)
geomean                                                                                                            1.540          1.398        -9.26%

                                                                                                                │  after3.txt  │             after4.txt              │
                                                                                                                │     B/op     │     B/op      vs base               │
JoinQuery/expr=rpc_request_success_total_+_rpc_request_error_total,steps=10000-10                                 54.53Mi ± 1%   54.08Mi ± 0%       ~ (p=0.481 n=10)
JoinQuery/expr=rpc_request_success_total_+_ON_(job,_instance)_GROUP_LEFT_rpc_request_error_total,steps=10000-10   1.245Gi ± 0%   1.245Gi ± 0%       ~ (p=0.143 n=10)
JoinQuery/expr=rpc_request_success_total_AND_rpc_request_error_total{instance=~"0.*"},steps=10000-10              38.18Mi ± 1%   38.42Mi ± 1%  +0.63% (p=0.019 n=10)
JoinQuery/expr=rpc_request_success_total_OR_rpc_request_error_total{instance=~"0.*"},steps=10000-10               38.58Mi ± 1%   38.80Mi ± 1%       ~ (p=0.393 n=10)
JoinQuery/expr=rpc_request_success_total_UNLESS_rpc_request_error_total{instance=~"0.*"},steps=10000-10           38.28Mi ± 1%   38.77Mi ± 1%  +1.27% (p=0.004 n=10)
geomean                                                                                                           82.92Mi        83.19Mi       +0.32%

                                                                                                                │ after3.txt  │             after4.txt             │
                                                                                                                │  allocs/op  │  allocs/op   vs base               │
JoinQuery/expr=rpc_request_success_total_+_rpc_request_error_total,steps=10000-10                                 562.8k ± 0%   561.8k ± 0%  -0.18% (p=0.001 n=10)
JoinQuery/expr=rpc_request_success_total_+_ON_(job,_instance)_GROUP_LEFT_rpc_request_error_total,steps=10000-10   20.57M ± 0%   20.57M ± 0%  -0.00% (p=0.000 n=10)
JoinQuery/expr=rpc_request_success_total_AND_rpc_request_error_total{instance=~"0.*"},steps=10000-10              307.5k ± 0%   310.1k ± 0%  +0.83% (p=0.000 n=10)
JoinQuery/expr=rpc_request_success_total_OR_rpc_request_error_total{instance=~"0.*"},steps=10000-10               307.5k ± 0%   310.1k ± 0%  +0.83% (p=0.001 n=10)
JoinQuery/expr=rpc_request_success_total_UNLESS_rpc_request_error_total{instance=~"0.*"},steps=10000-10           307.5k ± 0%   310.1k ± 0%  +0.83% (p=0.000 n=10)
geomean                                                                                                           804.3k        808.0k       +0.46%
```

#### Does this PR introduce a user-facing change?
```release-notes
NONE
```